fix(types): resolve the pyright errors of the package
Type fixes found by pyright in standard mode with the SciPy type stubs:

- cli: fail() is typed NoReturn, so the asserts after it are gone.
- rasterio windows are built with Window.from_slices, which type
  checkers can see, instead of the attrs constructor.
- Raster.resample validates scale and resolution in one branch chain;
  Raster.reproject reports a grid that rasterio could not compute.
- Normalization returns float32 as annotated; the class weight buffer of
  TaskLoss, the padding mode of Pad and the spline orders of the
  resampling functions carry declared types.
- component_statistics keeps its value statistics in one mapping and
  reads the centroids as an array; phase_unwrapping keeps the coherence
  optional.
- ONNX Runtime outputs, eigen-decompositions and kriging grids are
  converted to arrays explicitly.

+++ src/unbihexium/ai/transforms.py
@@ -102,7 +102,7 @@ class Normalization:
         # Means as float32 column vectors for broadcasting.
         self.mean = np.asarray(mean, dtype=np.float32).reshape(-1, 1, 1)
         # Standard deviations, bounded away from zero.
-        self.std = np.maximum(np.asarray(std, dtype=np.float32).reshape(-1, 1, 1), 1e-6)
+        self.std = np.maximum(np.asarray(std, dtype=np.float32).reshape(-1, 1, 1), np.float32(1e-6))
 
     # Number of bands.
     @property
@@ -169,7 +169,7 @@ class Normalization:
         # Standardise.
         x = (np.asarray(image, dtype=np.float32) - self.mean) / self.std
         # Replace missing values.
-        return np.nan_to_num(x, nan=0.0, posinf=0.0, neginf=0.0)
+        return np.nan_to_num(x, nan=0.0, posinf=0.0, neginf=0.0).astype(np.float32, copy=False)
 
     # Plain dictionary for checkpoints.
     def to_dict(self) -> dict[str, list[float]]:
